[jit] Split out Function into its own file
Some preparatory movement to prepare for the CU ownership overhaul.
Changes are purely mechanical.
