+++ aten/src/ATen/core/ivalue_inl.h
@@ -12,9 +12,7 @@
 
 namespace torch {
 namespace jit {
-namespace script {
 struct Function;
-}
 } // namespace jit
 } // namespace torch
 namespace c10 {
