+++ tools/build_variables.py
@@ -132,6 +132,7 @@ libtorch_sources = [
     "torch/csrc/jit/fuser/cpu/fused_kernel.cpp",
     "torch/csrc/jit/fuser/cpu/dynamic_library_unix.cpp",
     "torch/csrc/jit/fuser/interface.cpp",
+    "torch/csrc/jit/function.cpp",
     "test/cpp/jit/test.cpp",
 ]
