+++ torch/csrc/jit/ir.cpp
@@ -1429,7 +1429,7 @@ Node* Graph::createGetAttr(Value* obj, const std::string& field) {
 }
 
 Value* Graph::insertFunctionCall(
-    std::shared_ptr<script::Function> callee,
+    std::shared_ptr<Function> callee,
     script::MatchedSchema& matched) {
   Value* fn_constant = insertNode(create(prim::Constant))
                            ->output()
